Code cleanup complete
