+++ client/src/pages/Signup.jsx
@@ -1,100 +1,92 @@
+// this is the signup page, changes to the signup page can be made here
+
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-
 import { useMutation } from '@apollo/client';
-import { ADD_PROFILE } from '../utils/mutations';
-
+import { ADD_USER } from '../utils/mutations';
 import Auth from '../utils/auth';
 
-const Signup = () => {
-  const [formState, setFormState] = useState({
-    name: '',
-    email: '',
-    password: '',
-  });
-  const [addProfile, { error, data }] = useMutation(ADD_PROFILE);
-
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-
-    setFormState({
-      ...formState,
-      [name]: value,
+export default function Signup() { 
+    // set initial form state
+    const [formState, setFormState] = useState({ 
+        name: '',
+        email: '',
+        password: '' 
     });
-  };
+    // set state for form validation for the user generation functionality
+    const [addUser, { error, data }] = useMutation(ADD_USER);
 
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-    console.log(formState);
+    const handleChange = (event) => {
+        const { name, value } = event.target;
+        setFormState({ 
+            ...formState,
+            [name]: value 
+        });
+    };
 
-    try {
-      const { data } = await addProfile({
-        variables: { ...formState },
-      });
+    const handleFormSubmit = async (event) => { 
+        event.preventDefault();
+        try {
+            const { data } = await addUser({ 
+                variables: { ...formState } 
+            });
+            Auth.login(data.addUser.token);
+        } catch (e) {
+            console.error(e);
+        }
+    };
 
-      Auth.login(data.addProfile.token);
-    } catch (e) {
-      console.error(e);
-    }
-  };
-
-  return (
-    <main className="">
-      <section className="">
-        <section className="">
-          <h4 className="">Sign Up</h4>
-          <section className="">
-            {data ? (
-              <p>
-                Success! You may now head{' '}
-                <Link to="/">back to the homepage.</Link>
-              </p>
-            ) : (
-              <form onSubmit={handleFormSubmit}>
+    return (
+        <main className='container'> 
+        <section>
+            <section className='section-header'> 
+            <h2>Signup</h2>
+            </section>
+            <section className='section-body'>
+            { data ? (
+                <p>
+                Success! You may now head <Link to='/'>back to the homepage</Link>.
+                </p>
+            ):(
+                <form onSubmit={handleFormSubmit}>
                 <input
-                  className=""
-                  placeholder="Your username"
-                  name="name"
-                  type="text"
-                  value={formState.name}
-                  onChange={handleChange}
+                    className='form-input'
+                    placeholder='Your username'
+                    name='name'
+                    type='text'
+                    value={formState.name}
+                    onChange={handleChange}
                 />
                 <input
-                  className=""
-                  placeholder="Your email"
-                  name="email"
-                  type="email"
-                  value={formState.email}
-                  onChange={handleChange}
+                    className='form-input'
+                    placeholder='Your email'
+                    name='email'
+                    type='email'
+                    value={formState.email}
+                    onChange={handleChange}
                 />
                 <input
-                  className=""
-                  placeholder="******"
-                  name="password"
-                  type="password"
-                  value={formState.password}
-                  onChange={handleChange}
+                    className='form-input'
+                    placeholder='******'
+                    name='password'
+                    type='password'
+                    value={formState.password}
+                    onChange={handleChange}
                 />
-                <button
-                  className=""
-                  style={{ cursor: 'pointer' }}
-                  type="submit"
-                >
-                  Submit
+                <button className='button-link primary' type='submit'>
+                👽 Create User 👽
                 </button>
-              </form>
-            )}
+                </form>
 
+            )}
             {error && (
-              <section className="">
+                <section>
                 {error.message}
-              </section>
+                </section>
             )}
-          </section>
+            </section>
         </section>
-      </section>
-    </main>
-  );
-};
+        </main>
+    );
 
-export default Signup;
+};
