+++ client/src/utils/queries.js
@@ -1,28 +1,37 @@
+// this is the file that will hold the queries that will be used to make requests to the server to get data from the database.
+// queries are used to get data from the database using the graphQL server. 
+
 import { gql } from '@apollo/client';
 
-export const QUERY_PROFILES = gql`
-  query allProfiles {
-    profiles {
-      _id
-      name
-    }
-  }
-`;
-
-export const QUERY_SINGLE_PROFILE = gql`
-  query singleProfile($profileId: ID!) {
-    profile(profileId: $profileId) {
-      _id
-      name
-    }
-  }
-`;
-
+// this is the query that will be used to get the user's information for the logged in user
 export const QUERY_ME = gql`
-  query me {
-    me {
-      _id
-      name
+    query me {
+        me {
+            _id
+            name
+            email
+        }
     }
-  }
+`;
+
+// this is the query that will be used to get a user's profile information
+export const QUERY_USER = gql`
+    query profile($name: String!) {
+        profile(name: $name) {
+            _id
+            name
+            email
+        }
+    }
+`;
+
+// this is the query that will be used to get all the users in the database
+export const QUERY_USERS = gql`
+    query users {
+        users {
+            _id
+            name
+            email
+        }
+    }
 `;
