+++ server/models/index.js
@@ -1,7 +1,7 @@
 // this file is used to import all the models in the models folder and export them as an object for use in the express server.
 
 // import necessary models
-const Profile = require('./Profile'); 
+const User = require('./User'); 
 
 // export the models as an object for use in the express server
-module.exports = { Profile }; 
+module.exports = { User };
